Modify : 커스텀훅 분리, 파일 폴더화,변수명 수정
