+++ src/components/Nav/Dropdown/Carousel.scss
@@ -1,0 +1,46 @@
+.loopSection {
+  overflow: hidden;
+  margin: auto;
+}
+
+.carouselSection {
+  display: flex;
+  position: relative;
+  width: 400px;
+  height: 100%;
+  background-color: green;
+}
+
+.carousel {
+  display: flex;
+
+  .carouselImg {
+    width: 250px;
+  }
+
+  .carouselContent {
+    margin-top: 10%;
+
+    .carouselTitle {
+      width: 150px;
+      font-size: 24px;
+    }
+    .carouselText {
+      width: 150px;
+    }
+  }
+}
+
+.leftBtn {
+  position: absolute;
+  top: 50%;
+
+  left: 25%;
+}
+
+.rightBtn {
+  position: absolute;
+  top: 50%;
+
+  right: 25%;
+}
