+++ src/components/Nav/Dropdown/Dropdown.scss
@@ -1,0 +1,9 @@
+.dropDown {
+  position: absolute;
+  display: flex;
+  top: 60px;
+  left: 0px;
+  width: 100%;
+  height: 300px;
+  background-color: blue;
+}
