+++ src/components/Nav/Nav.scss
@@ -51,59 +51,3 @@
   padding-left: 10px;
   border-radius: 15px;
 }
-
-.dropDown {
-  position: absolute;
-  display: flex;
-  top: 60px;
-  left: 0px;
-  width: 100%;
-  height: 300px;
-  background-color: blue;
-}
-
-.loopSection {
-  overflow: hidden;
-  margin: auto;
-}
-
-.carouselSection {
-  display: flex;
-  position: relative;
-  width: 400px;
-  height: 100%;
-  background-color: green;
-}
-.carousel {
-  display: flex;
-
-  .carouselImg {
-    width: 250px;
-  }
-
-  .carouselContent {
-    margin-top: 10%;
-
-    .carouselTitle {
-      width: 150px;
-      font-size: 24px;
-    }
-    .carouselText {
-      width: 150px;
-    }
-  }
-}
-
-.leftBtn {
-  position: absolute;
-  top: 50%;
-
-  left: 25%;
-}
-
-.rightBtn {
-  position: absolute;
-  top: 50%;
-
-  right: 25%;
-}
