+++ src/components/Nav/Sideoption/Search/Searchresult.js
@@ -5,7 +5,7 @@ const Searchresult = props => {
 
   return (
     <div className="searchResult">
-      <img className="searchImg" src={img} alt="이미지" />
+      <img className="searchImg" src={img} alt="제품썸네일이미지" />
       <p>{value}</p>
       <p>{price}</p>
     </div>
